Separate tasks based on source of information
This commit changes the initial tasks and separates them based on the source. For apptweak, the response data is the same so we are able to create entities dynamically by giving it a type. For 42matters, we receive different fields and hence entity creation is done based on which app type we are quering.